Move Assign and Unassign Roles to SS.Client Auth DTO's
